# ui/public/openapi.yaml
openapi: 3.0.3
info:
  title: Evidra API
  description: |
    REST API for the Evidra reliability scoring platform.
    Receives evidence entries from CLI, MCP server, and CI pipelines.
    Computes behavioral signals and reliability scorecards.
  version: 1.0.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    url: https://github.com/vitas/evidra
servers:
  - url: /
    description: This server
security:
  - bearerAuth: []
...
paths:
...
  /v1/bench/leaderboard:
    get:
      summary: Model leaderboard (public, no auth)
      tags: [Bench]
      security: []
      parameters:
        - name: evidence_mode
          in: query
          schema:
            type: string
            enum: [all, none, evidra]
          description: |
            Top-level bench filter alias (All | Baseline | Evidra).
            Empty means all runs.
            `none` returns baseline runs only.
            `evidra` returns all non-`none` runs.
            Non-empty, non-`evidra` values are exact-match filters against stored modes.
        - name: k
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 3
          description: Number of trials for pass^k reliability metric
      responses:
        '200':
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  models:
                    type: array
                    items:
                      $ref: '#/components/schemas/LeaderboardEntry'
                  evidence_mode:
                    type: string
  /v1/bench/scenarios:
    get:
      summary: Scenario catalog
      tags: [Bench]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of scenarios
  /v1/bench/runs:
    get:
      summary: List benchmark runs
      tags: [Bench]
      parameters:
        - name: model
          in: query
          schema: { type: string }
        - name: scenario
          in: query
          schema: { type: string }
        - name: evidence_mode
          in: query
          schema: { type: string }
          description: |
            Empty means all runs.
            `none` returns baseline runs only.
            `evidra` returns all non-`none` runs.
            Non-empty, non-`evidra` values are exact-match filters against stored modes.
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: Paginated run list
    post:
      summary: Submit a benchmark run
      tags: [Bench]
      responses:
        '201':
          description: Run created
...
  /v1/bench/runs/archive:
    post:
      summary: Archive runs matching a filter
      tags: [Bench]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                before:
                  type: string
                  format: date-time
                  description: Archive runs created before this timestamp
                ids:
                  type: array
                  items: { type: string }
                  description: Archive runs with these IDs
                model:
                  type: string
                  description: Archive runs for this model
      responses:
        '200':
          description: Archive result
          content:
            application/json:
              schema:
                type: object
                properties:
                  archived:
                    type: integer
        '400':
          description: Missing filter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/bench/runs/{id}:
    get:
      summary: Get run detail
      tags: [Bench]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Run record
    delete:
      summary: Delete a single run
      tags: [Bench]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Run deleted
        '404':
          description: Run not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/bench/runs/{id}/transcript:
    get:
      summary: Run transcript (text/plain)
      tags: [Bench]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Transcript text
  /v1/bench/runs/{id}/tool-calls:
    get:
      summary: Tool call log (JSON)
      tags: [Bench]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Tool calls array
  /v1/bench/runs/{id}/timeline:
    get:
      summary: Decision timeline
      tags: [Bench]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Timeline with phases
  /v1/bench/runs/{id}/scorecard:
    get:
      summary: Run scorecard data
      tags: [Bench]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Scorecard JSON
  /v1/bench/stats:
    get:
      summary: Aggregate statistics
      tags: [Bench]
      parameters:
        - name: evidence_mode
          in: query
          schema:
            type: string
          description: |
            Top-level bench filter alias (All | Baseline | Evidra).
            Empty means all runs.
            `none` returns baseline runs only.
            `evidra` returns all non-`none` runs.
            Non-empty, non-`evidra` values are exact-match filters against stored modes.
      responses:
        '200':
          description: Stats
...
  /v1/bench/models:
    get:
      summary: List models enabled for the authenticated tenant
      tags: [Bench]
      responses:
        '200':
          description: Models available through platform defaults or tenant overrides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnabledModelsResponse'
        '500':
          description: Query failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/bench/models/{model_id}/provider:
    put:
      summary: Create or update a tenant-specific provider override for a model
      description: "Not yet enabled — requires AES-256-GCM key encryption."
      deprecated: true
      tags: [Bench]
      parameters:
        - name: model_id
          in: path
          required: true
          schema:
            type: string
          description: Bench model ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantProviderConfig'
            example:
              api_key: sk-secret
              api_base_url: https://gateway.example.com/v1
              rate_limit: 10
              monthly_budget: 100
      responses:
        '204':
          description: Tenant override saved
        '400':
          description: Invalid JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Update failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a tenant-specific provider override for a model
      description: "Not yet enabled — requires AES-256-GCM key encryption."
      deprecated: true
      tags: [Bench]
      parameters:
        - name: model_id
          in: path
          required: true
          schema:
            type: string
          description: Bench model ID
      responses:
        '204':
          description: Tenant override deleted
        '500':
          description: Delete failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v1/bench/compare/runs:
    get:
      summary: Compare two runs with computed delta
      tags: [Bench]
      parameters:
        - name: a
          in: query
          required: true
          schema: { type: string }
          description: First run ID
        - name: b
          in: query
          required: true
          schema: { type: string }
          description: Second run ID
      responses:
        '200':
          description: Run comparison with delta
  /v1/bench/compare/models:
    get:
      summary: Compare models across scenarios (pairwise or matrix)
      tags: [Bench]
      description: |
        Two modes:
        - Pairwise: ?a=X&b=Y — compare two models
        - Matrix: ?models=X,Y,Z&scenarios=A,B — multi-model grid (scenarios optional)
      parameters:
        - name: a
          in: query
          schema: { type: string }
          description: First model name (pairwise mode)
        - name: b
          in: query
          schema: { type: string }
          description: Second model name (pairwise mode)
        - name: models
          in: query
          schema: { type: string }
          description: Comma-separated model names (matrix mode)
        - name: scenarios
          in: query
          schema: { type: string }
          description: Comma-separated scenario IDs (matrix mode, optional)
        - name: evidence_mode
          in: query
          schema:
            type: string
          description: |
            Top-level bench filter alias (All | Baseline | Evidra).
            Empty means all runs.
            `none` returns baseline runs only.
            `evidra` returns all non-`none` runs.
            Non-empty, non-`evidra` values are exact-match filters against stored modes.
      responses:
        '200':
          description: Pairwise comparison or model matrix grid
  /v1/bench/signals:
    get:
      summary: Aggregated signal counts across runs
      tags: [Bench]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: evidence_mode
          in: query
          schema:
            type: string
          description: |
            Top-level bench filter alias (All | Baseline | Evidra).
            Empty means all runs.
            `none` returns baseline runs only.
            `evidra` returns all non-`none` runs.
            Non-empty, non-`evidra` values are exact-match filters against stored modes.
        - name: since
          in: query
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Signal aggregation
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_runs: { type: integer }
                  runs_with_scorecard: { type: integer }
                  signals:
                    type: object
                    additionalProperties:
                      type: object
                      properties:
                        total: { type: integer }
                        run_count: { type: integer }
                  avg_score: { type: number }
  /v1/bench/regressions:
    get:
      summary: Detect scenario/model regressions
      tags: [Bench]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: List of regressions (latest run failed, previous passed)
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    scenario_id: { type: string }
                    model: { type: string }
                    latest_run_id: { type: string }
                    latest_passed: { type: boolean }
                    prev_passed: { type: integer }
                    prev_total: { type: integer }
                    prev_rate: { type: number }
                    severity: { type: string, enum: [critical, warning] }
...
  /v1/bench/trigger/{id}:
    get:
      tags: [Bench]
      summary: Get trigger progress
      description: Returns current job state. Supports SSE streaming when Accept includes text/event-stream.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job state (JSON snapshot or SSE stream)
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  status: { type: string, enum: [pending, running, completed, failed] }
                  model: { type: string }
                  provider: { type: string }
                  total: { type: integer }
                  completed: { type: integer }
                  passed: { type: integer }
                  failed: { type: integer }
                  current_scenario: { type: string }
                  run_ids:
                    type: array
                    items: { type: string }
                  progress:
                    type: array
                    items:
                      type: object
                      properties:
                        scenario: { type: string }
                        status: { type: string, enum: [pending, running, passed, failed, error, skipped] }
                        run_id: { type: string }
                  created_at:
                    type: string
                    format: date-time
        '404':
          description: Job not found
...
  /v1/runners/{id}:
    delete:
      tags: [Bench]
      summary: Delete a registered runner
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Runner deleted
        '401':
          description: Unauthorized
        '404':
          description: Runner not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key issued via POST /v1/keys or EVIDRA_API_KEY env var
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    LeaderboardEntry:
      type: object
      properties:
        model:
          type: string
        scenarios:
          type: integer
        runs:
          type: integer
        pass_rate:
          type: number
          description: Run-weighted pass rate (0-100)
        avg_duration:
          type: number
        avg_cost:
          type: number
        total_cost:
          type: number
        pass_k:
          type: number
          description: "pass^k reliability — probability all k trials pass per scenario, averaged across qualifying scenarios (0-100)"
        pass_k_trials:
          type: integer
          description: k value used
        sufficient_scenarios:
          type: integer
          description: Scenarios with >= k trials
...